# openapi.yaml
openapi: 3.0.3
info:
  title: Fitlokal Clothing Store API
  description: |
    E-commerce backend for Fitlokal. Most routes require an **API key** (`x-api-key`) and many also require a **JWT** (`Authorization: Bearer <token>`).

    **Auth flow:** `POST /auth/register` or `POST /auth/login` → use returned `token` with `Authorization: Bearer ...` on protected routes.  
    **API key:** create or use an `applications` row (e.g. from seeders); send `x-api-key` on every request except pure `/auth/*` and public `/` / `robots.txt`.

    Admin-only actions require `users.isAdmin === true` on the JWT user.
  version: 1.0.0
servers:
  - url: "{baseUrl}"
    variables:
      baseUrl:
        default: http://localhost:3000
        description: API base URL (no trailing slash)
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Client application key from `applications.apiKey`
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from `POST /auth/login` or `POST /auth/register`
  schemas:
...
    ApiEnvelope:
      type: object
      properties:
        message:
          type: string
        data:
          type: object
          additionalProperties: true
      required: [message]
    ValidationError:
      type: object
      properties:
        message:
          type: string
          example: Validation error.
        data:
          type: object
          properties:
            error:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 64
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
        firstName:
          type: string
          maxLength: 128
          nullable: true
        lastName:
          type: string
          maxLength: 128
          nullable: true
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    AuthTokenResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
      example:
        message: Login successful
        data:
          token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
          user:
            id: 1
            username: shopper
            email: shopper@fitlokal.dev
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register user (email confirmation required)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered; returns token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "409":
          description: Email or username conflict
  /auth/register/confirm-email:
    get:
      tags: [Auth]
      summary: Confirm account email using verification token
      security: []
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Email confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "400":
          description: Validation error
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: JWT issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "401":
          description: Invalid credentials or disabled user
  /auth/login/magic-link:
    get:
      tags: [Auth]
      summary: Exchange magic-link JWT for session JWT
      security: []
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
        "400":
          description: Validation error
  /auth/login/magic-link/request:
    post:
      tags: [Auth]
      summary: Send a magic-link email via Resend
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Always returns success message to avoid email enumeration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
        "400":
          description: Validation error
...